+++ Cargo.toml
@@ -15,7 +15,7 @@ opt-level = 2
 [dependencies]
 serde = { version = "1", features = ["derive"] }
 eframe = { version = "0.28", default-features = false, features = [
-   #"accesskit",     # Make egui comptaible with screen readers. NOTE: adds a lot of dependencies.
+    #"accesskit",     # Make egui comptaible with screen readers. NOTE: adds a lot of dependencies.
     "default_fonts", # Embed the default egui fonts.
     "glow",          # Use the glow rendering backend. Alternative: "wgpu".
     "persistence",   # Enable restoring app state when restarting the app.
@@ -29,11 +29,16 @@ directories = "5"
 image = { version = "0.25", features = ["jpeg", "png"] }
 rand = "0.8"
 anyhow = "1.0"
+egui-notify = "0.15.0"
+once_cell = "1.19.0"
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.58.0", features = [
-    "Win32_UI_Shell", "Win32_Storage_FileSystem",
-    "Win32_System_Registry","Win32_System_Com","Win32_UI_WindowsAndMessaging"
+    "Win32_UI_Shell",
+    "Win32_Storage_FileSystem",
+    "Win32_System_Registry",
+    "Win32_System_Com",
+    "Win32_UI_WindowsAndMessaging",
 ] }
 
 [target.'cfg(windows)'.build-dependencies]
@@ -58,5 +63,5 @@ out_of_bounds_indexing = { level = "allow", priority = 7 }
 perf = { level = "warn", priority = 8 }
 style = { level = "warn", priority = 9 }
 # this should be active but for that anyhow should be used everywhere
-#unwrap_used = { level = "deny", priority = 10 }
+unwrap_used = { level = "deny", priority = 10 }
 #expect_used = { level = "deny", priority = 11 }
